got pixelized search working
Cleaned up and finalized pixelized search, seems to match naive calculation
as well as mangle in a couple tests. Realized in retrospect that it didn't
need to be as complex, but it does work.

Also cleaned up some interfaces.

Simplified the internal ra/dec to az/el conversion, which is different from
the "standard" but matches MANGLE's conversions.